refactor: rename directory with form for consistency
